web/cooking_flask/requirements.txt
flask
python-dotenv
pillow
pydantic
gunicorn